+++ GUI/__init__.py
@@ -1,11 +1,20 @@
+# Init Services
+from GUI.Services.course_adder import CourseAdder
+from GUI.Services.course_remover import CourseRemover
+from GUI.Services.course_updater import CourseUpdater
+from GUI.Services.filter_selecter import FilterSelecter
 from GUI.Services.user_authenticator import UserAuthenticator
 from GUI.Services.data_loader import DataLoader
 from GUI.Services.data_saver import DataSaver
 
+# Init Programs
 from GUI.prog_stat_analyzer import StatAnalyzer
 from GUI.prog_achievement_analyzer import AchievementAnalyzer
 from GUI.prog_grade_updater import GradeUpdater
 
+# Init Sub Frames
 from GUI.application import ApplicationFrame
 from GUI.login import LoginFrame
+
+# Init Main Frame
 from GUI.transcript_manager import TranscriptManager
